refactor(agent): extract trigger orchestration from task manager
